fix: audit fixes — typos, float credit formatting, referral reward in recovery
- Fix lstrip('@f') → lstrip('@') in support handler
- Fix 'Unknownf' → 'Unknown' and 'Nof' → 'No' string literals
- Fix state['phonef'] → state['phone'] KeyError
- Fix replace('+', 'f') → replace('+', '') in auth client names
- Fix verification.py log format %d → %s for float bonus
- Format verification.py reward notification with :g specifier
- Use fmt_credits() for all credit displays (menus, alerts, purchases, refunds)
- Change /addcred to parse float instead of int
- Add _check_referral_reward to main.py recovery processors
- Fix get_referral_earned return type hint to float
- Add /info admin command for user lookup
- Add "info" to command exclusion list in on_text handler

+++ database.py
@@ -466,7 +466,7 @@ async def get_referral_count(telegram_id: int) -> int:
     return await db.users.count_documents({"referred_by": telegram_id})
 
 
-async def get_referral_earned(telegram_id: int) -> int:
+async def get_referral_earned(telegram_id: int) -> float:
     user = await get_user(telegram_id)
     if not user:
         return 0
